+++ public/app/js/upload.js
@@ -7,7 +7,7 @@ $(document).ready(function() {
     // Upload immagine
    $(".insert-form > form").submit(function() {
         console.log('submit');
-        $(".loader").fadeOut("slow");
+        $(".loader").fadeIn("slow");
         return true;
     });