+++ CHANGELOG.md
@@ -6,7 +6,7 @@ TODO list
 - (MANUEL) modifica/cancellazione del immagine 
 - creazione/modifica/eliminazione del segnaposto per ogni immagine, anche integrazione con google-maps forse
 - (Eva, Nicola, Elisa Widmar, Alberto) validazione lato client del form di registrazione
-- () loader durante upload del immagine e associazione utente durante l'upload
+- (David) loader durante upload del immagine e associazione utente durante l'upload
 - lista filtrata per utente
 - login/logout e ricarica link sul login
 - migliorie UI/UX: loader, link dinamici